+++ build-scripts/python/build_documentation_toc.py
@@ -1,6 +1,8 @@
 """
 Build the table of contents for the online documentation, these are located under 'resources/documentation/<version>/'
 """
+from __future__ import annotations
+
 import warnings
 import time
 import json
@@ -9,6 +11,7 @@ import os
 
 from importlib import reload
 
+import pyfbsdk
 
 # Append current site-packages path
 CURRENT_DIR = os.path.dirname(__file__)
@@ -17,10 +20,12 @@ SITEPACKAGES_DIR = os.path.join(CURRENT_DIR, "site-packages")
 if SITEPACKAGES_DIR not in sys.path:
     sys.path.append(SITEPACKAGES_DIR)
 
+# pylint: disable=wrong-import-position
 import requests
 import js2py
 
 import pyfbsdk_stub_generator.plugins.online_documentation.documentation_scraper.table_of_contents as docScraper
+from pyfbsdk_stub_generator.plugins.online_documentation.documentation_scraper.page_parser import MemberItem
 reload(docScraper)
 
 
@@ -28,8 +33,21 @@ reload(docScraper)
 #            Structs & Enums
 # ------------------------------------------
 
-class FDictTags:
-    Url = "url"
+class EIcons:
+    """ 
+    VSCode icons (All icons can be found here: https://microsoft.github.io/vscode-codicons/dist/codicon.html)
+    """
+    FUNCTION = "symbol-function"
+    CLASS = "symbol-class"
+    PROPERTY = "symbol-property"
+    METHOD = "symbol-method"
+    ENUM = "symbol-enum"
+    ENUM_MEMBER = "symbol-enum-member"
+
+
+class EVsCodeItemData:
+    URL = "url"
+    LABEL = "label"
 
 
 # ------------------------------------------
@@ -38,7 +56,6 @@ class FDictTags:
 
 def get_motionbuilder_version():
     """ Get the current version of MotionBuilder """
-    import pyfbsdk
     return int(2000 + pyfbsdk.FBSystem().Version / 1000)
 
 
@@ -46,14 +63,16 @@ def get_output_directory():
     return os.path.join(CURRENT_DIR, "..", "..", "resources", "documentation", str(get_motionbuilder_version()))
 
 
-def save_json_file(filename, content):
+def save_items(filename, items: list):
     directory = get_output_directory()
     if not os.path.isdir(directory):
         os.makedirs(directory)
 
+    data = {"items": items}
+
     filepath = os.path.join(directory, filename)
     with open(filepath, "w+", encoding="utf-8") as file:
-        json.dump(content, file)
+        json.dump(data, file)
 
 
 # ------------------------------------------
@@ -62,33 +81,75 @@ def save_json_file(filename, content):
 
 def generate_examples_toc(version: int):
     examples_toc_url = f"https://help.autodesk.com/cloudhelp/{version}/ENU/MotionBuilder-SDK/py_ref/examples.js"
-    examples_response = requests.get(examples_toc_url, timeout = 10)
+    examples_response = requests.get(examples_toc_url, timeout=10)
     if examples_response.status_code != 200:
         # Log a warning:
-        warnings.warn(f"Failed to get examples table of contents from '{examples_toc_url}', status code: {examples_response.status_code}")
+        warnings.warn(
+            f"Failed to get examples table of contents from '{examples_toc_url}', status code: {examples_response.status_code}")
         return
 
     parsed_examples = js2py.eval_js(examples_response.text)
 
-    data = {}
+    items = []
     for title, url, children in parsed_examples:
-        data[title] = {FDictTags.Url: url}
+        items.append({
+            EVsCodeItemData.LABEL: title,
+            EVsCodeItemData.URL: url
+        })
 
-    save_json_file("examples.json", data)
+    save_items("examples.json", items)
 
 
 # ------------------------------------------
 #                Python
 # ------------------------------------------
 
+def is_enum(item: docScraper.TableOfContentItem):
+    if not hasattr(pyfbsdk, item.Name):
+        if item.Name == "Enumeration":
+            return True
+        return False
+
+    item_object = getattr(pyfbsdk, item.Name)
+    return type(item_object) == type  # pylint: disable=unidiomatic-typecheck
+
+
+def is_method(item: docScraper.TableOfContentItem, child: MemberItem):
+    if not hasattr(pyfbsdk, item.Name):
+        return False
+
+    item_object = getattr(pyfbsdk, item.Name)
+    if hasattr(item_object, child.Name):
+        child_object = getattr(item_object, child.Name)
+        return callable(child_object)
+
+    return False
+
+
 def generate_python_toc(version: int):
     python_docs = docScraper.Documentation("pyfbsdk", version, bUseCache=True)
 
-    data = {}
+    items = []
+    items_name_map = set()
     for item in python_docs.TableOfContents:
-        data[item.Name] = {FDictTags.Url: item.RelativeUrl}
-
         is_function = item.RelativeUrl.startswith("namespacepyfbsdk.html")
+
+        if is_function:
+            icon = EIcons.FUNCTION
+        elif is_enum(item):
+            icon = EIcons.ENUM
+        else:
+            icon = EIcons.CLASS
+
+        if item.Name in items_name_map:
+            continue
+
+        items.append({
+            EVsCodeItemData.LABEL: f"$({icon}) {item.Name}",
+            EVsCodeItemData.URL: item.RelativeUrl
+        })
+        items_name_map.add(item.Name)
+
         if is_function:
             continue
 
@@ -103,10 +164,25 @@ def generate_python_toc(version: int):
             if child.Name == item.Name:
                 continue
 
-            child_url = item.RelativeUrl + child.RelativeUrl
-            data[f"{item.Name}: {child.Name}"] = {FDictTags.Url: child_url}
+            title = f"{item.Name}: {child.Name}"
+            if title in items_name_map:
+                continue
+            items_name_map.add(title)
 
-    save_json_file("python.json", data)
+            if icon == EIcons.ENUM:
+                child_icon = EIcons.ENUM_MEMBER
+            elif is_method(item, child):
+                child_icon = EIcons.METHOD
+            else:
+                child_icon = EIcons.PROPERTY
+
+            child_url = item.RelativeUrl + child.RelativeUrl
+            items.append({
+                EVsCodeItemData.LABEL: f"$({child_icon}) {title}",
+                EVsCodeItemData.URL: child_url
+            })
+
+    save_items("python.json", items)
 
 
 # ------------------------------------------
@@ -122,7 +198,7 @@ def generate_table_of_contents(version):
         return delta_time
 
     _time_it(generate_examples_toc, version)
-    # _time_it(generate_python_toc, version)
+    _time_it(generate_python_toc, version)
 
 
 def main():